patch to convert mid-handshake stream into ssl stream
This is useful for async code where we expect accept() to result in
a mid-handshake stream and wish to have consistent types for all
accepted streams at the cost of managing handshaking state outside
of the type system.

This appears to save some unnecessary overheads in mio event-loop
driven code.